+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Revamp QR Scanner
 - Mobile: Move settings into bottom navigation
 - Enable Tether USDT for BTC Direct
+- Allow users to upgrade firmware during setup
 
 ## v4.51.4
 - Bundle BitBox02 and BitBox02 Nova firmware version v9.26.5